requirements.txt
# Core runtime dependencies
aiogram>=3.13,<4
aiogram-dialog>=2.2,<3
supabase>=2.0,<3
anthropic>=0.40,<1
httpx>=0.27,<1
pillow>=10.0,<11
python-dotenv>=1.0,<2
structlog>=24.0,<25
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
dateparser~=1.2
redis>=5.0,<6
croniter>=2.0,<7
